LayoutTests/storage/domstorage/localstorage/iframe-events-expected.txt

This is the main frame of a 2-frame document. Each frame is in the same security origin and therefore shares the same localStorage object. As a result, each frame should receive a StorageEvent when either frame changes the localStorage object.

Main frame about to run setItem on localStorage...
Main Frame received StorageEvent:
Key - Main Frame
New Value - SET
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/iframe-events.html
Storage Area - This window's window.localStorage

Subframe received storage event:
Key - Main Frame
New Value - SET
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/iframe-events.html
Storage Area - This window's window.localStorage

Subframe about to change localStorage...
Main Frame received StorageEvent:
Key - Subframe
New Value - SET
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/resources/iframe-events-second.html
Storage Area - This window's window.localStorage

Subframe received storage event:
Key - Subframe
New Value - SET
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/resources/iframe-events-second.html
Storage Area - This window's window.localStorage

Main frame about to run removeItem on localStorage...
Main Frame received StorageEvent:
Key - Main Frame
New Value - null
Old Value - SET
URI - LayoutTests/storage/domstorage/localstorage/iframe-events.html
Storage Area - This window's window.localStorage

Subframe received storage event:
Key - Main Frame
New Value - null
Old Value - SET
URI - LayoutTests/storage/domstorage/localstorage/iframe-events.html
Storage Area - This window's window.localStorage

Subframe about to change localStorage...
Main frame about to clear localStorage...
Main Frame received StorageEvent:
Key - 
New Value - null
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/iframe-events.html
Storage Area - This window's window.localStorage

Subframe received storage event:
Key - 
New Value - null
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/iframe-events.html
Storage Area - This window's window.localStorage

Subframe about to change localStorage...
Main Frame received StorageEvent:
Key - Subframe
New Value - SET
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/resources/iframe-events-second.html
Storage Area - This window's window.localStorage

Subframe received storage event:
Key - Subframe
New Value - SET
Old Value - null
URI - LayoutTests/storage/domstorage/localstorage/resources/iframe-events-second.html
Storage Area - This window's window.localStorage
